[Dask] Move resource methods to API
