+++ .importlinter
@@ -75,5 +75,4 @@ forbidden_modules=
     mlrun.api
 
 ignore_imports =
-    mlrun.runtimes.daskjob -> mlrun.api.utils.singletons.k8s
     mlrun.runtimes.utils -> mlrun.api.utils.singletons.k8s
