+++ mlrun/api/rundb/sqldb.py
@@ -692,6 +692,34 @@ class SQLRunDB(RunDBInterface):
             mask_params,
         )
 
+    def function_status(self, project, name, kind, selector):
+        """Retrieve status of a function being executed remotely (relevant to ``dask`` functions).
+
+        :param project:     The project of the function, not needed here.
+        :param name:        The name of the function, not needed here.
+        :param kind:        The kind of the function, currently ``dask`` is supported.
+        :param selector:    Selector clause to be applied to the Kubernetes status query to filter the results.
+        """
+        return self._transform_db_error(
+            mlrun.api.crud.Functions().get_function_status,
+            kind,
+            selector,
+        )
+
+    def start_function(
+        self, func_url: str = None, function: "mlrun.runtimes.BaseRuntime" = None
+    ):
+        """Execute a function remotely, Used for ``dask`` functions.
+
+        :param func_url: URL to the function to be executed.
+        :param function: The function object to start.
+        :returns: A BackgroundTask object, with details on execution process and its status.
+        """
+        return self._transform_db_error(
+            mlrun.api.crud.Functions().start_function,
+            function,
+        )
+
     def list_pipelines(
         self,
         project: str,
