+++ mlrun/api/runtime_handlers/daskjob.py
@@ -21,14 +21,24 @@ from sqlalchemy.orm import Session
 import mlrun.common.schemas
 import mlrun.errors
 import mlrun.k8s_utils
+import mlrun.runtimes
+import mlrun.runtimes.pod
 import mlrun.utils
 import mlrun.utils.regex
 from mlrun.api.db.base import DBInterface
 from mlrun.api.runtime_handlers.base import BaseRuntimeHandler
 from mlrun.config import config
 from mlrun.runtimes.base import RuntimeClassMode
-from mlrun.runtimes.utils import get_k8s
+from mlrun.runtimes.utils import get_k8s, get_resource_labels
 from mlrun.utils import logger
+
+
+def get_dask_resource():
+    return {
+        "scope": "function",
+        "start": deploy_function,
+        "status": get_obj_status,
+    }
 
 
 class DaskRuntimeHandler(BaseRuntimeHandler):
@@ -207,3 +217,188 @@ class DaskRuntimeHandler(BaseRuntimeHandler):
                 # ignore error if service is already removed
                 if exc.status != 404:
                     raise
+
+
+def deploy_function(
+    function: mlrun.runtimes.DaskCluster,
+    secrets=None,
+    client_version: str = None,
+    client_python_version: str = None,
+):
+    _validate_dask_related_libraries_installed()
+
+    scheduler_pod, worker_pod, function, namespace = enrich_dask_cluster(
+        function, secrets, client_version, client_python_version
+    )
+    return initialize_dask_cluster(scheduler_pod, worker_pod, function, namespace)
+
+
+def initialize_dask_cluster(scheduler_pod, worker_pod, function, namespace):
+    import dask
+    import dask_kubernetes
+
+    spec, meta = function.spec, function.metadata
+
+    svc_temp = dask.config.get("kubernetes.scheduler-service-template")
+    if spec.service_type or spec.node_port:
+        if spec.node_port:
+            spec.service_type = "NodePort"
+            svc_temp["spec"]["ports"][1]["nodePort"] = spec.node_port
+        mlrun.utils.update_in(svc_temp, "spec.type", spec.service_type)
+
+    norm_name = mlrun.utils.normalize_name(meta.name)
+    dask.config.set(
+        {
+            "kubernetes.scheduler-service-template": svc_temp,
+            "kubernetes.name": "mlrun-" + norm_name + "-{uuid}",
+        }
+    )
+
+    cluster = dask_kubernetes.KubeCluster(
+        worker_pod,
+        scheduler_pod_template=scheduler_pod,
+        deploy_mode="remote",
+        namespace=namespace,
+        idle_timeout=spec.scheduler_timeout,
+    )
+
+    logger.info(f"cluster {cluster.name} started at {cluster.scheduler_address}")
+
+    function.status.scheduler_address = cluster.scheduler_address
+    function.status.cluster_name = cluster.name
+    if spec.service_type == "NodePort":
+        ports = cluster.scheduler.service.spec.ports
+        function.status.node_ports = {
+            "scheduler": ports[0].node_port,
+            "dashboard": ports[1].node_port,
+        }
+
+    if spec.replicas:
+        cluster.scale(spec.replicas)
+    else:
+        cluster.adapt(minimum=spec.min_replicas, maximum=spec.max_replicas)
+
+    return cluster
+
+
+def enrich_dask_cluster(
+    function, secrets, client_version: str = None, client_python_version: str = None
+):
+    from dask.distributed import Client, default_client  # noqa: F401
+    from dask_kubernetes import KubeCluster, make_pod_spec  # noqa: F401
+    from kubernetes import client
+
+    # Is it possible that the function will not have a project at this point?
+    if function.metadata.project:
+        function._add_secrets_to_spec_before_running(project=function.metadata.project)
+
+    spec = function.spec
+    meta = function.metadata
+    spec.remote = True
+
+    image = (
+        function.full_image_path(
+            client_version=client_version, client_python_version=client_python_version
+        )
+        # TODO: we might never enter here, since running a function requires defining an image
+        or "daskdev/dask:latest"
+    )
+    env = spec.env
+    namespace = meta.namespace or config.namespace
+    if spec.extra_pip:
+        env.append(spec.extra_pip)
+
+    pod_labels = get_resource_labels(function, scrape_metrics=config.scrape_metrics)
+    # TODO: 'dask-worker' is deprecated, new dask CLI was introduced in 2022.10.0.
+    #  Upgrade when we drop python 3.7 support and use 'dask worker' instead
+    worker_args = ["dask-worker", "--nthreads", str(spec.nthreads)]
+    memory_limit = spec.worker_resources.get("limits", {}).get("memory")
+    if memory_limit:
+        worker_args.extend(["--memory-limit", str(memory_limit)])
+    if spec.args:
+        worker_args.extend(spec.args)
+    # TODO: 'dask-scheduler' is deprecated, new dask CLI was introduced in 2022.10.0.
+    #  Upgrade when we drop python 3.7 support and use 'dask scheduler' instead
+    scheduler_args = ["dask-scheduler"]
+
+    container_kwargs = {
+        "name": "base",
+        "image": image,
+        "env": env,
+        "image_pull_policy": spec.image_pull_policy,
+        "volume_mounts": spec.volume_mounts,
+    }
+    scheduler_container = client.V1Container(
+        resources=spec.scheduler_resources, args=scheduler_args, **container_kwargs
+    )
+    worker_container = client.V1Container(
+        resources=spec.worker_resources, args=worker_args, **container_kwargs
+    )
+
+    scheduler_pod_spec = mlrun.runtimes.pod.kube_resource_spec_to_pod_spec(
+        spec, scheduler_container
+    )
+    worker_pod_spec = mlrun.runtimes.pod.kube_resource_spec_to_pod_spec(
+        spec, worker_container
+    )
+    for pod_spec in [scheduler_pod_spec, worker_pod_spec]:
+        if spec.image_pull_secret:
+            pod_spec.image_pull_secrets = [
+                client.V1LocalObjectReference(name=spec.image_pull_secret)
+            ]
+
+    scheduler_pod = client.V1Pod(
+        metadata=client.V1ObjectMeta(namespace=namespace, labels=pod_labels),
+        # annotations=meta.annotation),
+        spec=scheduler_pod_spec,
+    )
+    worker_pod = client.V1Pod(
+        metadata=client.V1ObjectMeta(namespace=namespace, labels=pod_labels),
+        # annotations=meta.annotation),
+        spec=worker_pod_spec,
+    )
+    return scheduler_pod, worker_pod, function, namespace
+
+
+def _validate_dask_related_libraries_installed():
+    try:
+        import dask  # noqa: F401
+        from dask.distributed import Client, default_client  # noqa: F401
+        from dask_kubernetes import KubeCluster, make_pod_spec  # noqa: F401
+        from kubernetes import client  # noqa: F401
+    except ImportError as exc:
+        print(
+            "missing dask or dask_kubernetes, please run "
+            '"pip install dask distributed dask_kubernetes", %s',
+            exc,
+        )
+        raise exc
+
+
+def get_obj_status(selector=None, namespace=None):
+    if selector is None:
+        selector = []
+
+    import mlrun.api.utils.singletons.k8s
+
+    k8s = mlrun.api.utils.singletons.k8s.get_k8s_helper()
+    namespace = namespace or config.namespace
+    selector = ",".join(["dask.org/component=scheduler"] + selector)
+    pods = k8s.list_pods(namespace, selector=selector)
+    status = ""
+    for pod in pods:
+        status = pod.status.phase.lower()
+        if status == "running":
+            cluster = pod.metadata.labels.get("dask.org/cluster-name")
+            logger.info(
+                "Found running dask function",
+                pod_name=pod.metadata.name,
+                cluster=cluster,
+            )
+            return status
+        logger.info(
+            "Found dask function in non ready state",
+            pod_name=pod.metadata.name,
+            status=status,
+        )
+    return status
